youtube video checkbox and input reset
I added a checkbox and border for an upcoming option to bring up youtube so you can work alongside a youtube video. Also made it so the input box resets on add task button press. For some reason you can't use pre-defined variables and you have to rewrite the id from the doc and set the value to '';